build: Change Docker base image to ruby:X [LARA-5]
This changes the base image to no longer use the private base image and instead lifts the RailsLTS password into the LARA configuration.

The reason for this is the number of different Ruby upgrades that are required coupled with the required updates to the gems to support that Ruby version.  This will be easier to upgrade using a direct base image.