feat: Added object storage config to runtime/report init data [LARA-199]
This commit adds an "opaque" object storage configuration object to the
initInteractive data for both the runtime and report APIs.

This config is opaque to the interactive, and is intended to be passed
through to the @concord-consortium/object-storage
library to configure object storage access for interactives.

The question-interactives interactives will automatically make use of
this config when it is present, at the base runtime level so that each
interactive does not need to handle it individually.  Interactives outside
of question-interactives can also make use of this config by passing it
to the @concord-consortium/object-storage library.

The @concord-consortium/object-storage defines the concrete type for
the object storage config.  Currently there is a "demo" config that
stores objects in memory only.  A "firebase" config will be added soon
to store objects in a Firebase via the Activity Player supplied
credentials.