feat: Add pubsub to LARA api [LARA-205]
This commit adds a publish-subscribe (pubsub) system to the LARA
interactive API. This allows interactives to create channels, publish
messages to those channels, and subscribe/unsubscribe to channels to
receive messages.

A new PubSubManager class is added to the interactive-api-host to manage
channels and subscriptions. This class will be used by hosts like AP
to handle pubsub requests from interactives.

The interactive-api-client is updated with
new methods to create channels, publish messages, and manage
subscriptions.  This allows interactives to easily use the pubsub system.